refactor playground component
