+++ src/Playground.js
@@ -1,11 +1,11 @@
-import { useState, useEffect, useRef, createContext, useContext, useCallback, useLayoutEffect } from 'react';
-import execute from './executor';
+import { useState, useEffect, useRef, useLayoutEffect } from 'react';
 import { uploadPsycho } from './executor';
 import { useSettings } from './hooks/useSettings';
-import Editor, { useMonaco } from "@monaco-editor/react"
-import { registerPHPSnippetLanguage } from "./utils/registerPHPSnippetLanguage";
 import Split from 'split.js'
-import { PlaygroundProvider, useLoading } from './contexts/PlaygroundContext';
+import { PlaygroundProvider } from './contexts/PlaygroundContext';
+import Input from './playground/Input';
+import Output from './playground/Output';
+import LoadingIndicator from './LoadingIndicator';
 
 const editorOptions = {
     lineHeight: 32,
@@ -26,8 +26,7 @@ const editorOptions = {
     }
 }
 
-
-function Playground({ project }) {
+export default function Playground({ project }) {
     const [output, setOutput] = useState('')
     const [settings,] = useSettings()
     const splitInstance = useRef(null)
@@ -59,11 +58,11 @@ function Playground({ project }) {
                     backgroundColor: 'rgb(23, 23, 23)'
                 }}>
                 <div className="pg-input overflow-hidden">
-                    <Input setOutput={setOutput} project={project} />
+                    <Input {...{ setOutput, project, editorOptions }} />
                 </div>
 
                 <div className="pg-output overflow-hidden relative">
-                    <Output output={output} />
+                    <Output {...{ output, editorOptions }} />
                     <LoadingIndicator />
                 </div>
 
@@ -71,87 +70,3 @@ function Playground({ project }) {
         </PlaygroundProvider>
     )
 }
-
-function Input({ setOutput, project }) {
-    const [, setLoading] = useLoading()
-    const [code,] = useState("// Press Ctr/Cmd + Enter to run code \necho 'Welcome to Tinker 2'")
-    const monaco = useMonaco()
-    let editorRef = useRef(null);
-
-    const runCode = useCallback(() => {
-        setLoading(true)
-        execute({ code: editorRef.current.getValue(), project })
-            .then(({ stdout }) => {
-                try {
-                    const result = JSON.parse(stdout.trim())
-                    setOutput(result.output)
-                } catch (ex) {
-                    setOutput(stdout.trim())
-                }
-            })
-            .finally(() => setLoading(false))
-    }, [project, setLoading, setOutput])
-
-    useEffect(() => {
-        if (monaco) {
-            registerPHPSnippetLanguage(monaco.languages)
-
-            if (editorRef.current) {
-                editorRef.current.addCommand(monaco.KeyMod.CtrlCmd | monaco.KeyCode.Enter, runCode)
-            }
-        }
-    }, [monaco, runCode])
-
-    function handleEditorDidMount(editor, monaco) {
-        editorRef.current = editor;
-        editorRef.current.addCommand(monaco.KeyMod.CtrlCmd | monaco.KeyCode.Enter, runCode)
-    }
-
-    return (
-        <Editor
-            key="tinker-pad"
-            theme="vs-dark"
-            language="php-snippet"
-            value={code}
-            // onChange={(value) => setCode(value)}
-            onMount={handleEditorDidMount}
-            options={editorOptions}
-        />
-    )
-}
-
-function Output({ output }) {
-    return (
-        <>
-            <div className="h-full">
-                <Editor
-                    key="output"
-                    language="php-snippet"
-                    theme="vs-dark"
-                    value={output}
-                    options={{
-                        ...editorOptions,
-                        readOnly: true,
-                        wordWrap: 'on',
-                        renderIndentGuides: false,
-                        contextmenu: false,
-                    }}
-                />
-            </div>
-        </>
-    )
-}
-
-function LoadingIndicator() {
-    const [loading,] = useLoading()
-    const loadingClass = loading ? 'animate-spin' : 'hidden';
-    return (
-        <div className="absolute top-2 right-2 text-white">
-            <svg xmlns="http://www.w3.org/2000/svg" className={`h-5 w-5 ` + loadingClass} viewBox="0 0 20 20" fill="currentColor">
-                <path fillRule="evenodd" d="M4 2a1 1 0 011 1v2.101a7.002 7.002 0 0111.601 2.566 1 1 0 11-1.885.666A5.002 5.002 0 005.999 7H9a1 1 0 010 2H4a1 1 0 01-1-1V3a1 1 0 011-1zm.008 9.057a1 1 0 011.276.61A5.002 5.002 0 0014.001 13H11a1 1 0 110-2h5a1 1 0 011 1v5a1 1 0 11-2 0v-2.101a7.002 7.002 0 01-11.601-2.566 1 1 0 01.61-1.276z" clipRule="evenodd" />
-            </svg>
-        </div>
-    )
-}
-
-export default Playground;
